Stop job on control-C (not other exceptions) and carry on to report results
The previous behaviour was:

* On any exception, the job would be stopped.
* ...and no results would be reported to Jenkins (no jUnit xml file
  created).

In other words, the previous behaviour was:

* Transient network errors (RequestsException) caused the job to be
  stopped and no results to be reported to Jenkins.
* Control-C caused the job to be stopped and no results to be reported
  to Jenkins.

With this commit, the behaviour is:

* Transient network errors cause the script to exit immediately. Still
  no results will be reported to Jenkins, but at least the test-job will
  continue to run and you can see the results in the Stb-tester Portal.
  We will implement retries for transient network problems in a future
  commit (see https://github.com/stb-tester/stbt-rig/pull/19).

* Control-C (or SIGTERM) does stop the job as before, but now it also
  creates the jUnit results xml file for Jenkins.

I had previously implemented this in commit 772bc56 (pull request #20)
but in an incredibly convoluted way.

I'm not sure what the correct behaviour of `run_tests(...,
await_completion=True)` should be when used as a **Python API** so I
have left that unchanged. This commit only applies to usage as a
command-line program.