pytest plugin: Fix paths when run from test-pack root
`File.name` was the file basename; `File.fspath` is the absolute path of
the file.

This is a step towards supporting VS Code's test discovery & running.

I assume it worked in PyCharm because PyCharm runs pytest from the
directory containing the file. VS Code runs pytest from the project
root.