pytest plugin: Disable pytest's built-in python plugin
We use stbt_rig as a pytest plugin to integrate with the "run this test"
functionality of some IDEs. Previously (for PyCharm) we told people to
disable pytest's built-in python plugin with the command line option
"-p no:python" (so that pytest doesn't try to run the test locally on
the user's PC).

However, now I'm trying to use the same pytest integration with VS Code,
but VS Code uses pytest for test *discovery* too, and it specifically
expects the `pytest.Function` class from pytest (not our
`StbtRemoteTest` class). Duck-typing won't help us here because VS Code
specifically checks for the `pytest.Function` type.

As VS Code ignores our `StbtRemoteTest` type, we can return both
`StbtRemoteTest` *and* the built-in `pytest.Function` during test
collection. We can tell we're being invoked in VS Code's test discovery
phase because of the "--collect-only" argument to pytest.

Another option would have been to always disable pytest's built-in
python plugin, and make our `StbtRemoteTest` class inherit from
`pytest.Function` instead of `pytest.Item` (note that `pytest.Function`
is a subclass of `pytest.Item`). But then we'd have to implement a more
sophisticated collector that actually imports the Python files. Maybe
it wouldn't be so hard, if we make our collector inherit from pytest's
`PyCollector` and we override `collect`.