Use refs/snapshots/$username instead of refs/heads/$username/snapshot
This way the snapshot ref isn't a normal branch (because it's in
`refs/snapshots` not `refs/heads`) which used to confuse some
users (e.g. you could create a pull request from the snapshot branch,
but on your next save it'd overwrite the contents of that branch without
preserving any history of the previous snapshots).

Note: We need to call `portal.notify_push` because github only notifies
for normal branches in `refs/heads`. We only need to call
`portal.notify_push` from the explicit snapshot command; when we call
`push_git_snapshot` from `run_tests`, we immediately follow up by
telling the portal to run the tests, and we tell the portal what the new
SHA is so the portal will do a `git fetch` if it doesn't have the SHA.