snapshot: Use file locking to prevent simultaneous snapshots
If we have multiple `stbt_rig.py snapshot`s running simultaneously we can
get errors on push where github responds with "Rejected".  This can happen
particularly when doing a "Save" and then clicking "Run tests" in VSCode.
It gives users confusing errors.

This avoids that issue by using file locking for inter-process
synchronisation.  The lock is held over both the commit and the push parts
of the process so we don't end up pushing an old snapshot.