Ask again if you enter a blank access token
I was testing this in a new VM, and when I pasted `py stbt_rig.py setup
--vscode` into my VS Code terminal it took a while at the beginning
without any output; I wasn't sure if it was running so I pressed RETURN
again. Much later when it came to ask for the token, it processed that
RETURN, printed this, and exited:

    EOF!
    Traceback (most recent call last):
    [...]
      File "C:\Users\IEUser\stb-tester-test-pack-demo\stbt_rig.py", line 903, in __init__
        self._auth_token = next(self._it)
    StopIteration

I looked into flushing stdin before we call `input` (note that ask =
input). But that seems to be extremely complicated.[^1]
I can't do `sys.stdin.read()` because it blocks until you close stdin
with Control-D.
`sys.stdin.flush()` does nothing because flush is a no-op on read-only
streams.
I can't do `sys.stdin.seek(0, io.SEEK_END)` because stdin is not
seekable.

[^1]: https://stackoverflow.com/questions/21791621/taking-input-from-sys-stdin-non-blocking