setup --vscode: Configure git.fetchOnPull and git.pruneOnFetch
So that when I press the "Synchronize Changes" button in the status bar,
it fetches new branches (particularly, I want it to fetch branches
created by the Object Repository on the Stb-tester Portal). Without
this, VS Code only pull the current branch, and we'd have to teach the
user the difference between "git fetch" and VS Code's "Synchronize
Changes" button.

Also prune deleted remote branches, so that the user doesn't have to do
an explicit "Fetch (prune)" which is buried deep in VS Code's "Source
Control" tab > triple-dot menu > "Pull, Push" > "Fetch (Prune)".

> Git: Fetch On Pull
> When enabled, fetch all branches when pulling. Otherwise, fetch just
> the current one.
>
> Git: Prune On Fetch
> Prune when fetching.