vscode: Use pylint binary from venv
So that we control the pylint & astroid version, for consistent results.

This forces the extension to shell out to pylint. Without this, the
extension tries to import pylint instead of running it as a subprocess,
and if we use "pylint.importStrategy: fromEnvironment" to import our
version of pylint, that causes all sorts of compatibility issues because
the extension's dependencies (like `attrs`) come from our venv instead
of the extension.